app: Update project to use iio_app from iio/iio_app
This fixes broken project.
